Move Lingvo estimator to speech recognition module
Furthermore, remove the sequence mixin abstract class definition.

All ASR-related estimators are now part of the `speech_recognition`
module.
