--- art/estimators/sequence/__init__.py
@@ -1,2 +1,0 @@
-from art.estimators.sequence.sequence import SequenceNetworkMixin
-from art.estimators.sequence.tensorflow import LingvoAsr
