+++ art/estimators/speech_recognition/__init__.py
@@ -4,3 +4,4 @@ Module containing estimators for speech recognition.
 from art.estimators.speech_recognition.speech_recognizer import SpeechRecognizerMixin
 
 from art.estimators.speech_recognition.pytorch_deep_speech import PyTorchDeepSpeech
+from art.estimators.speech_recognition.tensorflow_lingvo import LingvoAsr
